Module-ize some vlsirtools.netlist
